Factor WinRT GetFolderPath into static accessible via reflection
